ptu/requirements_pipreqs.txt
matplotlib==3.10.8
numpy==2.4.1
opencv_python==4.12.0.88
Pillow==12.1.0
pyfiglet==1.0.4
schedule==1.2.2
streamlit==1.50.0
termcolor==3.3.0
ultralytics==8.4.0
